changed folder sturcture and fixed tests
