+++ pom.xml
@@ -8,15 +8,17 @@
 		<version>2.7.0</version>
 		<relativePath/> <!-- lookup parent from repository -->
 	</parent>
-	<groupId>com.crossword-service</groupId>
-	<artifactId>crossword-service</artifactId>
+	<groupId>de.unistuttgart</groupId>
+	<artifactId>crossword-backend</artifactId>
 	<version>0.0.1-SNAPSHOT</version>
-	<name>crossword-service</name>
-	<description>Crossword service</description>
+	<name>Crossword-puzzle backend</name>
+	<description>Backend for the crossword-puzzle</description>
 	<properties>
 		<java.version>17</java.version>
         <lombok.version>1.18.24</lombok.version>
         <org.mapstruct.version>1.5.2.Final</org.mapstruct.version>
+        <plugin.prettier.goal>write</plugin.prettier.goal>
+        <org.springdoc.version>1.6.9</org.springdoc.version>
 	</properties>
 	<dependencies>
 		<dependency>
@@ -28,6 +30,11 @@
             <artifactId>lombok</artifactId>
             <version>${lombok.version}</version>
             <scope>provided</scope>
+        </dependency>
+        <dependency>
+            <groupId>org.springdoc</groupId>
+            <artifactId>springdoc-openapi-ui</artifactId>
+            <version>${org.springdoc.version}</version>
         </dependency>
         <dependency>
             <groupId>org.mapstruct</groupId>
@@ -82,7 +89,33 @@
                     </execution>
                 </executions>
             </plugin>
-
+            <plugin>
+                <groupId>com.hubspot.maven.plugins</groupId>
+                <artifactId>prettier-maven-plugin</artifactId>
+                <version>0.16</version>
+                <configuration>
+                    <prettierJavaVersion>1.5.0</prettierJavaVersion>
+                    <printWidth>120</printWidth>
+                    <tabWidth>2</tabWidth>
+                    <useTabs>false</useTabs>
+                    <ignoreConfigFile>true</ignoreConfigFile>
+                    <ignoreEditorConfig>true</ignoreEditorConfig>
+                    <!-- Use <inputGlobs> to override the default input patterns -->
+                    <inputGlobs>
+                        <!-- These are the default patterns, you can omit <inputGlobs> entirely unless you want to override them -->
+                        <inputGlob>src/main/java/**/*.java</inputGlob>
+                        <inputGlob>src/test/java/**/*.java</inputGlob>
+                    </inputGlobs>
+                </configuration>
+                <executions>
+                    <execution>
+                        <phase>validate</phase>
+                        <goals>
+                            <goal>${plugin.prettier.goal}</goal>
+                        </goals>
+                    </execution>
+                </executions>
+            </plugin>
 			<plugin>
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-jar-plugin</artifactId>
@@ -90,7 +123,7 @@
                 <configuration>
                     <archive>
                         <manifest>
-                            <mainClass>com.crosswordservice.CrosswordServiceApplication</mainClass>
+                            <mainClass>de.unistuttgart.crosswordbackend.CrosswordServiceApplication</mainClass>
                         </manifest>
                     </archive>
                 </configuration>
