+++ src/main/java/de/unistuttgart/crosswordbackend/CrosswordServiceApplication.java
@@ -1,4 +1,4 @@
-package com.crosswordservice;
+package de.unistuttgart.crosswordbackend;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
