+++ src/main/java/de/unistuttgart/crosswordbackend/CrosswordServiceConfiguration.java
@@ -1,4 +1,4 @@
-package com.crosswordservice;
+package de.unistuttgart.crosswordbackend;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.CorsRegistry;
