+++ src/main/java/de/unistuttgart/crosswordbackend/crosswordchecker/CrosswordChecker.java
@@ -1,7 +1,7 @@
-package com.crosswordservice.crosswordchecker;
+package de.unistuttgart.crosswordbackend.crosswordchecker;
 
-import com.crosswordservice.data.Crossword;
-import com.crosswordservice.data.Question;
+import de.unistuttgart.crosswordbackend.data.Crossword;
+import de.unistuttgart.crosswordbackend.data.Question;
 import lombok.NoArgsConstructor;
 
 import java.util.List;
