+++ src/main/java/de/unistuttgart/crosswordbackend/repositories/QuestionRepository.java
@@ -1,13 +1,9 @@
-package com.crosswordservice.repositories;
+package de.unistuttgart.crosswordbackend.repositories;
 
-import com.crosswordservice.data.Question;
+import de.unistuttgart.crosswordbackend.data.Question;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Modifying;
-import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
-import javax.transaction.Transactional;
-import java.util.List;
 import java.util.UUID;
 
 /**
